pyproject.toml
[tool.poetry]
name = "goit-module11"
version = "0.1.0"
description = ""
authors = ["romanlenger <145775490+romanlenger@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pathlib = "^1.0.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-mail = "^1.4.2"
fastapi-cache2 = "^0.2.2"
aioredis = "^2.0.1"
redis = "^5.2.1"
setuptools = "^75.6.0"
cloudinary = "^1.41.0"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
